Chore (./contracts): Linting

+++ contracts/src/verifiers/DepositVerifier.sol
@@ -22,17 +22,17 @@ pragma solidity >=0.7.0 <0.9.0;
 
 contract Groth16Verifier {
     // Scalar field size
-    uint256 constant r    = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
+    uint256 constant r = 21888242871839275222246405745257275088548364400416034343698204186575808495617;
     // Base field size
-    uint256 constant q   = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
+    uint256 constant q = 21888242871839275222246405745257275088696311157297823662689037894645226208583;
 
     // Verification Key data
-    uint256 constant alphax  = 7551512397760434695278787407970089361920976085724591893631715113665513493155;
-    uint256 constant alphay  = 4858743534958594702299513812002532330030483289133666508947444335896973309544;
-    uint256 constant betax1  = 7138136986218441300139150885539646496770234652739011163448551918795965653539;
-    uint256 constant betax2  = 4712825566973883170357105386159418790555493841853209642048622776393876364613;
-    uint256 constant betay1  = 19907689548832711461939838245098301347723498620225312171842325865135980934261;
-    uint256 constant betay2  = 6113291132796547763696096414016747469354975217717077088790776678006800547003;
+    uint256 constant alphax = 7551512397760434695278787407970089361920976085724591893631715113665513493155;
+    uint256 constant alphay = 4858743534958594702299513812002532330030483289133666508947444335896973309544;
+    uint256 constant betax1 = 7138136986218441300139150885539646496770234652739011163448551918795965653539;
+    uint256 constant betax2 = 4712825566973883170357105386159418790555493841853209642048622776393876364613;
+    uint256 constant betay1 = 19907689548832711461939838245098301347723498620225312171842325865135980934261;
+    uint256 constant betay2 = 6113291132796547763696096414016747469354975217717077088790776678006800547003;
     uint256 constant gammax1 = 11559732032986387107991004021392285783925812861821192530917403151452391805634;
     uint256 constant gammax2 = 10857046999023057135944570762232829481370756359578518086990519993285655852781;
     uint256 constant gammay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
@@ -42,30 +42,33 @@ contract Groth16Verifier {
     uint256 constant deltay1 = 4082367875863433681332203403145435568316851327593401208105741076214120093531;
     uint256 constant deltay2 = 8495653923123431417604973247489272438418190587263600148770280649306958101930;
 
-    
     uint256 constant IC0x = 10528101766570989236532259563501793387246198596274239845756475009134314907238;
     uint256 constant IC0y = 9785310478589352646338183923889914337258119906362965275999425476795000161768;
-    
+
     uint256 constant IC1x = 20524348950630442655807040893677928827144580351547034872693251634582565310942;
     uint256 constant IC1y = 14626378207935328725564292420297815438897218555759224295844050021539110315641;
-    
+
     uint256 constant IC2x = 20569135388978928413237771483609523317264428256478485136619804649081866235044;
     uint256 constant IC2y = 17733752525823702481654894736863860254839968399346818313160755732675776973770;
-    
+
     uint256 constant IC3x = 2414945871568083223355862064140564548002758732125210960045182263778402332809;
     uint256 constant IC3y = 10321953642258548666427589861907486903367669585463741889167579460299561101406;
-    
+
     uint256 constant IC4x = 13172308763953719591749157952807462444706511923094942298427397701272522705483;
     uint256 constant IC4y = 14373543888432929627621958191007449578550402601679105897476356286624696982540;
-    
- 
+
     // Memory data
     uint16 constant pVk = 0;
     uint16 constant pPairing = 128;
 
     uint16 constant pLastMem = 896;
 
-    function verifyProof(uint[2] calldata _pA, uint[2][2] calldata _pB, uint[2] calldata _pC, uint[4] calldata _pubSignals) public view returns (bool) {
+    function verifyProof(
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[4] calldata _pubSignals
+    ) public view returns (bool) {
         assembly {
             function checkField(v) {
                 if iszero(lt(v, q)) {
@@ -73,7 +76,7 @@ contract Groth16Verifier {
                     return(0, 0x20)
                 }
             }
-            
+
             // G1 function to multiply a G1 value(x,y) to value in an address
             function g1_mulAccC(pR, x, y, s) {
                 let success
@@ -108,15 +111,14 @@ contract Groth16Verifier {
                 mstore(add(_pVk, 32), IC0y)
 
                 // Compute the linear combination vk_x
-                
+
                 g1_mulAccC(_pVk, IC1x, IC1y, calldataload(add(pubSignals, 0)))
-                
+
                 g1_mulAccC(_pVk, IC2x, IC2y, calldataload(add(pubSignals, 32)))
-                
+
                 g1_mulAccC(_pVk, IC3x, IC3y, calldataload(add(pubSignals, 64)))
-                
+
                 g1_mulAccC(_pVk, IC4x, IC4y, calldataload(add(pubSignals, 96)))
-                
 
                 // -A
                 mstore(_pPairing, calldataload(pA))
@@ -142,7 +144,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 384), mload(add(pMem, pVk)))
                 mstore(add(_pPairing, 416), mload(add(pMem, add(pVk, 32))))
 
-
                 // gamma2
                 mstore(add(_pPairing, 448), gammax1)
                 mstore(add(_pPairing, 480), gammax2)
@@ -159,7 +160,6 @@ contract Groth16Verifier {
                 mstore(add(_pPairing, 704), deltay1)
                 mstore(add(_pPairing, 736), deltay2)
 
-
                 let success := staticcall(sub(gas(), 2000), 8, _pPairing, 768, _pPairing, 0x20)
 
                 isOk := and(success, mload(_pPairing))
@@ -169,23 +169,22 @@ contract Groth16Verifier {
             mstore(0x40, add(pMem, pLastMem))
 
             // Validate that all evaluations ∈ F
-            
+
             checkField(calldataload(add(_pubSignals, 0)))
-            
+
             checkField(calldataload(add(_pubSignals, 32)))
-            
+
             checkField(calldataload(add(_pubSignals, 64)))
-            
+
             checkField(calldataload(add(_pubSignals, 96)))
-            
+
             checkField(calldataload(add(_pubSignals, 128)))
-            
 
             // Validate all evaluations
             let isValid := checkPairing(_pA, _pB, _pC, _pubSignals, pMem)
 
             mstore(0, isValid)
-             return(0, 0x20)
-         }
-     }
- }
+            return(0, 0x20)
+        }
+    }
+}
